Apply theme in app for consistancy
